refactor: update tool documentation and consolidate resource paths
