+++ tests/tools/_spyclient.py
