+++ tests/tools/test_vocabulary.py
